requirements.txt
instagrapi
google-generativeai
flask
